Update script and introduce init.d script
This commit updates LibreSwan to version 3.13. This is a critical
fix, as all versions priort to 3.13 are susceptible to a IKE
daemon restart attack, which could make DoS attacks very easy.

Furthermore the rc.local instructions are ditched. Instead the
script is handled by an init.d script. This has the advantage
that it's more stable (or more clean at least) and it makes it
easier to restart the VPN.
WARNING! When you rerun the script to update, make sure to
reset or delete your old rc.local.

The random password is now longer, due to security concerns.

The backup process of configuration files is smoother.

The setup echoes are much more informative on what files are
being modified.

This update also fixes some formatting issues.