Update script for other systems + add IP checking
In this commit:

Updated the script for:
- Arch Linux
- CentOS
- Fedora

Added a check for the version of Internet Protocol to be used, IPv4 or IPv6.

Why this check?
Most servers nowadays already have IPv6 and therefore when connection to the IP address check site (wtfismyip.com) is made with IPv6, causing the IPv6 address being used when configuring the VPN server.

When a user with only IPv4 connectivity tries to connect to a server configured that way, the server is reachable but cannot be established a connection with.

Now, with the option of configuring with IPv4/IPv6, choices can be made. If the user decides that his connection and the websites he/she would be using would support IPv6, then he can use that :)

Otherwise, it would be a wiser choice to choose IPv4, since most people are still using it.

Signed-off-by: Edwin Ang <edwin@theroyalstudent.com>